requirements.txt
pose_format
tqdm~=4.63.1
numpy~=1.22.3
pylint
pytest
wandb==0.12.11
pytorch_lightning==1.6.0
torch==1.11.0
torchmetrics~=0.7.3
opencv-python==4.5.5.64
mediapipe
pympi-ling
